Fix: created new components for dialogs
