+++ apps/delivery/delivery/src/app/delivery/form/form.component.ts
@@ -1,41 +1,8 @@
-import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
-import { TemplateQuery, TemplateService } from '../../template/+state';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { TemplateQuery } from '../../template/+state';
 import { Observable } from 'rxjs';
-import { MatDialogRef, MatDialog } from '@angular/material';
-
-@Component({
-  selector: 'new-template-dialog',
-  template: `
-    <div mat-dialog-content>
-      <h2>Save as a new template</h2>
-      <mat-form-field appearance="outline" fxLayoutAlign="center">
-        <mat-label>File name</mat-label>
-        <input #templateName matInput placeholder="New template name" />
-      </mat-form-field>
-    </div>
-    <div mat-dialog-actions fxLayoutAlign="center">
-      <button mat-stroked-button color="primary" (click)="close()">Cancel</button>
-      <button mat-raised-button color="primary" (click)="saveTemplate(templateName.value)">
-        Save Template
-      </button>
-    </div>
-  `
-})
-export class NewTemplateDialogComponent {
-  constructor(
-    public dialogRef: MatDialogRef<NewTemplateDialogComponent>,
-    private templateService: TemplateService
-  ) {}
-
-  public saveTemplate(templateName) {
-    this.templateService.saveTemplate(templateName);
-    this.dialogRef.close();
-  }
-
-  public close() {
-    this.dialogRef.close();
-  }
-}
+import { MatDialog } from '@angular/material';
+import { NewTemplateComponent } from './new-template.component';
 
 @Component({
   selector: 'delivery-form',
@@ -56,6 +23,6 @@ export class FormComponent implements OnInit {
   }
 
   public openDialog() {
-    this.dialog.open(NewTemplateDialogComponent);
+    this.dialog.open(NewTemplateComponent);
   }
 }
